Execute SQL script (#37)
* Add script to execute SQL
* Add copy script

We need to be able to execute the DDL schema against the database, ideally without someone connecting to the production database from their local machine and doing it by hand.

This script automates the running of the schema defined as *.sql files, and runs it on application boot, the same way that Sequelize and other ORMs do.

Since the SQL files will be read alphabetically, I renamed all of them to be prepended with a 2 digit 0# number that indicates the order they should be executed in. This is similar to how knex.js generates filenames with the current timestamp when creating a new migration file.