Add seed script (#38)
This seeds the PayID service with a single payment pointer of https://xpring.io/hansbergren, which resolves with the following on a GET request:

{
  "xAddress": "T7WSFgh6owANWoD2V3WRg6aeBveBzExkpDowirvnLDGL2YW",
  "classicAddress": "rsGDoRqqefkvWtznKe7bHH6HtmjnJwQvWK"
}

This is the last step needed for a GET request demo for the PayID service.

Now, if the wallet is configured with a proxy that sends https://xpring.io/{:username} to the PayID service. Everything should "Just Work™".

Note, once the demo is over, we should definitely remove this.