Use transactions for multiple database statements (#120)
While the golden path for inserting or updating a user was fine before, the error cases were not.

Since inserting into the Account table and inserting into the Address table were two separate operations, it was possible to insert a user into the Account table without inserting the addresses associated with that user. If that ever happened, we would have sent a 500 - Internal Server Error to the client, but they would have actually inserted a user into the database! This would have led to our data being inconsistent/unpredictable.

Now, we use knex.transaction() to ensure that the data in our database is always consistent.

In particular:
- Remove insertAddresses/updateAddresses because that functionality is now rolled into insertUser/updateUser.
- Rename selectAddresses to selectUser, to match the other DB CRUD functions.
- Disable ESLint's no-use-before-define rule for helper functions across the codebase.
- Allow unused variables / parameters when destructuring an array/object, since that's an elegant way to partially clone an object.
- Wrap insertUser and updateUser in a DB transaction to allow commit/rollback of the entire data change.
- Add two helper functions to reduce duplication between insertUser and updateUser.