Add a location header and relevant tests (#151)
201 - Created responses should include a Location header in the response that tells the client where the created resource now lives: #139

This implements Location headers for 201 responses from POST and PUT, as well as adding assertions around the Location header in E2E tests with a 201 response.

Context of Change:
I chose to implement this by adding the paymentPointer to res.locals, and then generating the Location header programmatically in the shared.ts middleware that sends out our final responses.

This is because POST and PUT both need this, and I didn't want any duplicated logic. I also chose to get our /v1/users/ URL out of req.originalUrl. This way, we don't need to import config just for the current version number, and if the /users/ path ever changes, we'll pick it up automatically.