Use Log4JS for logging (#185)
Replaces console.log and console.error calls with a custom logger using the log4js library.

The two biggest disadvantages of console.log are that:
- You can't easily toggle it on and off
- There is no real notion of "log levels"

This PR introduces the log4js dependency to use for logging. This lets us remove existing configuration around logging (which we disable in our tests to reduce noise).