Rename some files (#232)
I was unhappy with some of the filenames / file locations.

Basically, it was weird that shared.ts only had one export, and that it was sendSuccess. I kept looking for sendSuccess.ts, but it didn't exist.

Also, I wanted to separate out knex SQL queries from other business logic in the /services directory. You can imagine in the future that we might still have a /services/users.ts, but now it's business logic & validation, instead of SQL queries.

In particular:

shared.ts => sendSuccess.ts
/services/users.ts => /data-access/users.ts
/services/payIds.ts => /data-access/payIds.ts