[2/x] Adds getAllPaymentInfoFromDatabase (#242)
This PR adds `getAllPaymentInfoFromDatabase`, which fetches all of the payment information associated with a `payId`. I noticed that there was a type called `AddressInfo` but decided it was better to use a type specific to `data-access/payIds` to decouple `Address`.

Note that this PR **does not** use this function; this will be used in the future to sort through the results to find the best available requested type for content-negotiation.