[3/x] Handles multiple Accept types (#243)
This PR implements content-negotiation and tests to verify that it works.

In order to keep the actual DB query simple, and because we aren't supporting that many network/environment pairs, I decided to do the sort at the app level. Since I imagine this is a...rarer use case, I also tried to preserve the current behavior of requesting a single network/environment pair.

In future PRs, I want to do the sort in the DB, which will simplify all of the application logic. An intermediate version could include a data access function called something like `getPaymentInformationForNetworks` or whatever