Remove organization column (#296)
This gets rid of the organization column in the account table, since we don't use it at all.

Initially, we were building this out to be a "hosted service" where you can imagine multiple different organizations using the same "PayID provider". However, for that to work, we would need some sort of authentication on the private API, to identify which organization was making that request.

Since we decided against implementing any private API authentication, I think it makes sense to rip out the organization column, since it was hardcoded in the code to 'xpring' anyways.

Also, now that we have Prometheus metrics, there's an PAYID_ORG environment variable that indicates which organization is running the PayID server, so when we collect metrics we know who its from. So, removing the organization column removes a potential point of confusion.

In particular, this PR:

- Adds a migration strategy
  - New migrations directory for SQL files, and run it when we run - 
  - syncDatabaseSchema
- Removes the organization column
  - Migration script
  - Remove from canonical table definition account.sql file
  - Remove from types/database.ts
  - Remove organization references in Knex queries
  - Update seed script to not insert into the organization column.