Fix error payloads for 409-Conflict (#293)
This is an issue I caught in one of my linting PRs. Basically we weren't checking the error message / error response payload that got returned when we hit a 409 - Conflict.

This was a little unusual because we do check the error response payloads for all our other error codes.

We actually had a bug, where the HTTP response code was correctly 409, but the response body said it was a 500, because we didn't handle the 409 case in handleHttpError.