Private API version (#418)
Adds a version header to the private API, to provide a better experience for any partner that uses our reference implementation as their production PayID server.

The PayID-API-Version header needs to be an ISO8601 date string, of the form YYYY-MM-DD. This is what Coinbase, Stripe, and various other companies use to version their hosted API services.

Note that I kept the /v1/ in the path for the private API. I chose to do this as Coinbase, Stripe, etc. do the same thing. We should never need to increment to /v2/ in the path, but it's cheap to leave it there, makes a smaller diff, and gives us a safety valve for the future.