Add issue templates (#468)
This adds issue templates to our repo, in preparation of open-sourcing.

Bug Report
Feature Request
Security Issue (which really just tells you to email us and not open an issue).

This is largely based on ESLint's GitHub setup. You can see what this looks like when you open a new issue on ESLint.