Add Liche for CI (#562)
Fixes broken links in the index.html file you get when you hit a PayID server directly from a browser. Reported in #560 .

Concretely, we update the "Getting Started" link to point to the new Getting Started endpoint on the docs website, and we update the "Developer Docs" link to point to the PayID Overview section of the docs website.

Also, add [liche](https://github.com/raviqqe/liche) to our CI so broken links won't occur in the future.