Add a report for count of PayIDs (#594)
Adds a report for the count of the number of PayIDs.

Also rename the existing report logic to reflect that it is actually a count of addresses.

We currently only capture a count of addresses grouped by (paymentNetwork, environment), but it can be misleading if you think that is a count of unique PayIDs. Now, we have a separate Prometheus Gauge that counts up the number of unique PayIDs per server.